backend/requirements.txt
fastapi>=0.94.0,<0.99.0
pydantic>=1.8.0,<2.0.0
uvicorn==0.27.1
sqlalchemy==1.4.25
psycopg2-binary==2.9.6    #linux/mac user use: psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-jose
passlib==1.7.4
pydantic<2.0.0
PyJWT
python-decouple
python-multipart
requests
starlette>=0.27.0,<0.28.0
pytest>=6.0.0<7.0.0
httpx >= 0.20.0
coverage
pytest-asyncio
pylint
pytest-cov==5.0.0
locust